some last little bits
