+++ src/backend/apiUtils/workspaces/index.ts
@@ -6,9 +6,8 @@ import { getOrCreateStripeCustomer } from '../stripe/customers';
 
 import { defaultWorkspaceLabels } from './defaultLabels';
 
-import { AuthDecoratedRequest } from 'types/auth';
+import { AuthDecoratedRequest, ExtendedSession } from 'types/auth';
 import prisma from 'backend/prisma/client';
-import { ExtendedSession } from 'types/common';
 import { ForbiddenException } from 'backend/apiUtils/exceptions';
 import { PermissionType } from 'types/permissions';
