+++ src/types/auth/index.ts
@@ -1,7 +1,18 @@
 import { NextApiRequest } from 'next';
+import { Session } from 'next-auth';
 
-import { ExtendedSession } from '../common/index';
+import { EnrichedWorkspaceMember } from 'types/common';
 
 export interface AuthDecoratedRequest extends NextApiRequest {
   session: ExtendedSession;
 }
+
+export interface ExtendedToken {
+  email: string;
+  name: string;
+  picture: string;
+  workspaceMemberships: EnrichedWorkspaceMember[];
+  sub: string;
+}
+
+export type ExtendedSession = Session & { token: ExtendedToken };
